Use a single message editor in employee-frontend
